refactor(apprefresh): split refresh.go by domain + encapsulate session model cache (F5 + F15)
F5: refresh.go shrinks from 1293 → 323 lines. parseOpenclawConfig (~400 lines) split into per-section helpers (parseCompaction, parseSkills, parseModelDefaults, parseAvailableModels, parseChannels, parseHooks, parsePlugins, parseSkillsCfg, parseBindings, parseModelParams, parseAvailModelsList, parseAgents) under refresh_config.go. Other functions moved to dedicated files preserving wire format:
- refresh_gateway.go: collectGatewayHealth
- refresh_git.go: collectGitLog
- refresh_alerts.go: BuildAlerts, BuildCostBreakdown, statusError/statusOffline constants
- refresh_chart.go: BuildDailyChart (with frozen-merge contract comment)
- refresh_crons.go: loadCronStateSidecar, CollectCrons
- refresh_helpers.go: jsonObj/jsonArr/jsonStr/jsonStrDefault/asObj/aliasOrID/round2/sumBucketCosts/FilterByDate/LimitSlice/ensureMapMap/ensureMapMapInt/sortedJSONKeys

refresh.go now contains only RunRefreshCollector, the orchestrator collectDashboardData, and label helpers (TitleCase/TrimLabel/ModelName).

Also: data.json perm tightened to 0o600 (was 0o644) — bundles aggregate cost figures, no need for world-readable (F22 pulled from Wave 4).

F15: liveSessionModelCache becomes a real struct with injectable fetch/runner/resolver fields. defaultSessionModelCache singleton preserves existing call-site contract via the package-level getLiveSessionModels wrapper. New session_model_cache_test.go demonstrates per-instance isolation (t.Parallel-safe) and pins the singleflight invariant under contention.

All existing tests pass with -race.
